refactor: 언론사 리스트 파일 구조 수정
